Add files via upload
sudoku.py now iterates through the image in each cell and tells if there is a number written in it. If there is, the image gets passed into the CNN for predicting digits. The entire image matrix is stored in 'sudokuMatrix.' If a cell does not have a number, I gave it the value 10.